arrange hierarchy
